requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
redis
langchain-google-genai
langgraph
langchain-core
langchain
loguru
pydantic
python-dotenv
sentry-sdk
httpx
pytest
aiohttp==3.9.1
websockets==12.0
duckduckgo-search
langchain-community
langchain-tavily
amadeus 
serpapi
googlemaps